Fix AA Cancel Desyncs
Adds the packet NPC_Instant_Stop_Attack from LeaguePackets and replaces all previous uses of NotifyStopAutoAttack with NotifyInstantStopAttack.
NOTE: Missile Net IDs should be fixed in the future, as when using them in LeaguePackets' packets, they do not work properly. Because of this, the missileNetID parameter of NotifyInstantStopAttack defaults to 0.