Fix the Rest of Vision
Fixed sometimes losing vision of a unit when a raycast stops where it started.

Fixed not being able to see through the NavGridFlag 0x40 (SEE_THROUGH, or TransparentWall), this consequently fixes not being able to attack units that are behind a see through part of the NavGrid, such as Inhibitors and Nexus'.
NOTE: This is only applicable if the NavGrid has the flag.
In the case were it does not, unless you decide to edit the NavGrid, use the fix suggested by https://github.com/LeagueSandbox/GameServer/pull/845